Remove substrate-consensus-common dependency from core-client
