+++ core/consensus/aura/src/lib.rs
@@ -59,7 +59,8 @@ use std::time::{Duration, Instant};
 use codec::Encode;
 use consensus_common::{Authorities, BlockImport, Environment, Proposer};
 use client::ChainHead;
-use consensus_common::{ImportBlock, BlockOrigin};
+use client::client::BlockOrigin;
+use consensus_common::ImportBlock;
 use runtime_primitives::{generic, generic::BlockId};
 use runtime_primitives::traits::{Block, Header, Digest, DigestItemFor};
 use network::import_queue::{Verifier, BasicQueue};
