+++ core/network/src/test/sync.rs
@@ -17,7 +17,7 @@
 use client::backend::Backend;
 use client::blockchain::HeaderBackend as BlockchainHeaderBackend;
 use config::Roles;
-use consensus::BlockOrigin;
+use client::client::BlockOrigin;
 use sync::SyncState;
 use super::*;
